Working!

+++ vertex.h
@@ -3,9 +3,9 @@
 
 struct vertex3d_t
 {
-    int x;
-    int y;
-    int z;
+    int16_t x;
+    int16_t y;
+    int16_t z;
 };
 
 #endif /* VERTEX_H */
